initial support for two versions of Kolasu
