+++ src/interop/ecore.ts
@@ -12,274 +12,24 @@ import {Point, Position} from "../position";
 import {Parser} from "../parsing";
 import {Parser as ANTLRParser, ParserRuleContext} from "antlr4ts";
 import {Issue, IssueSeverity, IssueType} from "../validation";
-
-// Kolasu model definition
+import {getEPackage} from "./ecore_basic";
 
 export const TO_EOBJECT_SYMBOL = Symbol("toEObject");
 export const ECLASS_SYMBOL = Symbol("EClass");
 export const EPACKAGE_SYMBOL = Symbol("EPackage");
 export const SYMBOL_NODE_NAME = Symbol("name");
 
-export const KOLASU_URI_V1 = "https://strumenta.com/kolasu/v1";
-export const KOLASU_URI_V2 = "https://strumenta.com/kolasu/v2";
-export const THE_AST_RESOURCE = Ecore.ResourceSet.create().create({ uri: KOLASU_URI_V2 });
-export const THE_AST_EPACKAGE = getEPackage("com.strumenta.kolasu.v2", { nsURI: KOLASU_URI_V2 });
-THE_AST_RESOURCE.get("contents").add(THE_AST_EPACKAGE);
-export const THE_ORIGIN_ECLASS = Ecore.EClass.create({
-    name: "Origin",
-    abstract: true
-});
-export const THE_NODE_ECLASS = Ecore.EClass.create({
-    name: "ASTNode",
-    abstract: true,
-});
-THE_NODE_ECLASS.get("eSuperTypes").add(THE_ORIGIN_ECLASS);
-export const THE_POINT_ECLASS = Ecore.EClass.create({
-    name: "Point"
-});
-THE_POINT_ECLASS.get("eStructuralFeatures").add(Ecore.EAttribute.create({
-    name: "line",
-    eType: Ecore.EInt,
-    lowerBound: 1
-}));
-THE_POINT_ECLASS.get("eStructuralFeatures").add(Ecore.EAttribute.create({
-    name: "column",
-    eType: Ecore.EInt,
-    lowerBound: 1
-}));
-export const THE_POSITION_ECLASS = Ecore.EClass.create({
-    name: "Position"
-});
-THE_POSITION_ECLASS.get("eStructuralFeatures").add(Ecore.EReference.create({
-    name: "start",
-    eType: THE_POINT_ECLASS,
-    containment: true,
-    lowerBound: 1
-}));
-THE_POSITION_ECLASS.get("eStructuralFeatures").add(Ecore.EReference.create({
-    name: "end",
-    eType: THE_POINT_ECLASS,
-    containment: true,
-    lowerBound: 1
-}));
-THE_NODE_ECLASS.get("eStructuralFeatures").add(Ecore.EReference.create({
-    name: "position",
-    eType: THE_POSITION_ECLASS,
-    containment: true
-}));
-THE_NODE_ECLASS.get("eStructuralFeatures").add(Ecore.EReference.create({
-    name: "destination",
-    eType: THE_POSITION_ECLASS,
-    containment: true,
-    lowerBound: 0
-}));
-THE_NODE_ECLASS.get("eStructuralFeatures").add(Ecore.EReference.create({
-    name: "origin",
-    eType: THE_ORIGIN_ECLASS,
-    containment: false,
-    lowerBound: 0
-}));
-
-export const THE_POSSIBLY_NAMED_INTERFACE = Ecore.EClass.create({
-    name: "PossiblyNamed",
-    interface: true
-});
-THE_POSSIBLY_NAMED_INTERFACE.get("eStructuralFeatures").add(Ecore.EAttribute.create({
-    name: "name",
-    eType: Ecore.EString,
-    lowerBound: 0
-}));
-export const THE_NAMED_INTERFACE = Ecore.EClass.create({
-    name: "Named",
-    interface: true
-});
-THE_NAMED_INTERFACE.get("eSuperTypes").add(THE_POSSIBLY_NAMED_INTERFACE);
-THE_NAMED_INTERFACE.get("eStructuralFeatures").add(Ecore.EAttribute.create({
-    name: "name",
-    eType: Ecore.EString,
-    lowerBound: 1
-}));
-export const THE_REFERENCE_BY_NAME_ECLASS = Ecore.EClass.create({
-    name: "ReferenceByName"
-});
-THE_REFERENCE_BY_NAME_ECLASS.get("eSuperTypes").add(THE_NAMED_INTERFACE);
-THE_REFERENCE_BY_NAME_ECLASS.get("eTypeParameters").add(Ecore.ETypeParameter.create({
-    name: "N"
-}));
-THE_REFERENCE_BY_NAME_ECLASS.get("eTypeParameters").at(0).get("eBounds").add(Ecore.EGenericType.create({
-    eClassifier: THE_POSSIBLY_NAMED_INTERFACE
-}));
-THE_REFERENCE_BY_NAME_ECLASS.get("eStructuralFeatures").add(Ecore.EAttribute.create({
-    name: "name",
-    eType: Ecore.EString,
-    lowerBound: 1
-}));
-THE_REFERENCE_BY_NAME_ECLASS.get("eStructuralFeatures").add(Ecore.EReference.create({
-    name: "referenced",
-    containment: true
-}));
-THE_REFERENCE_BY_NAME_ECLASS.get("eStructuralFeatures").at(1).set("eGenericType", Ecore.EGenericType.create({
-    eTypeParameter: THE_REFERENCE_BY_NAME_ECLASS.get("eTypeParameters").at(0)
-}));
-
-export const THE_LOCAL_DATE_ECLASS  = Ecore.EClass.create({
-    name: "LocalDate"
-});
-THE_LOCAL_DATE_ECLASS.get("eStructuralFeatures").add(Ecore.EAttribute.create({
-    name: "year",
-    eType: Ecore.EInt,
-    lowerBound: 1
-}));
-THE_LOCAL_DATE_ECLASS.get("eStructuralFeatures").add(Ecore.EAttribute.create({
-    name: "month",
-    eType: Ecore.EInt,
-    lowerBound: 1
-}));
-THE_LOCAL_DATE_ECLASS.get("eStructuralFeatures").add(Ecore.EAttribute.create({
-    name: "dayOfMonth",
-    eType: Ecore.EInt,
-    lowerBound: 1
-}));
-
-export const THE_LOCAL_TIME_ECLASS  = Ecore.EClass.create({
-    name: "LocalTime"
-});
-THE_LOCAL_TIME_ECLASS.get("eStructuralFeatures").add(Ecore.EAttribute.create({
-    name: "hour",
-    eType: Ecore.EInt,
-    lowerBound: 1
-}));
-THE_LOCAL_TIME_ECLASS.get("eStructuralFeatures").add(Ecore.EAttribute.create({
-    name: "minute",
-    eType: Ecore.EInt,
-    lowerBound: 1
-}));
-THE_LOCAL_TIME_ECLASS.get("eStructuralFeatures").add(Ecore.EAttribute.create({
-    name: "second",
-    eType: Ecore.EInt,
-    lowerBound: 1
-}));
-THE_LOCAL_TIME_ECLASS.get("eStructuralFeatures").add(Ecore.EAttribute.create({
-    name: "nanosecond",
-    eType: Ecore.EInt,
-    lowerBound: 1
-}));
-
-export const THE_LOCAL_DATE_TIME_ECLASS  = Ecore.EClass.create({
-    name: "LocalDateTime"
-});
-THE_LOCAL_DATE_TIME_ECLASS.get("eStructuralFeatures").add(Ecore.EReference.create({
-    name: "date",
-    eType: THE_LOCAL_DATE_ECLASS,
-    lowerBound: 1,
-    containment: true
-}));
-THE_LOCAL_DATE_TIME_ECLASS.get("eStructuralFeatures").add(Ecore.EReference.create({
-    name: "time",
-    eType: THE_LOCAL_TIME_ECLASS,
-    lowerBound: 1,
-    containment: true
-}));
-
-
-function addLiteral(eenum: Ecore.EEnum, name: string, value: number) {
-    const literal = Ecore.EEnumLiteral.create({
-        name, value
-    });
-    eenum.get("eLiterals").add(literal);
-    return literal;
-}
-
-export const THE_ISSUE_TYPE_EENUM = Ecore.EEnum.create({
-    name: "IssueType"
-});
-addLiteral(THE_ISSUE_TYPE_EENUM, IssueType[IssueType.LEXICAL], IssueType.LEXICAL);
-addLiteral(THE_ISSUE_TYPE_EENUM, IssueType[IssueType.SYNTACTIC], IssueType.SYNTACTIC);
-addLiteral(THE_ISSUE_TYPE_EENUM, IssueType[IssueType.SEMANTIC], IssueType.SEMANTIC);
-
-export const THE_ISSUE_SEVERITY_EENUM = Ecore.EEnum.create({
-    name: "IssueSeverity"
-});
-addLiteral(THE_ISSUE_SEVERITY_EENUM, IssueSeverity[IssueSeverity.INFO], IssueSeverity.INFO);
-addLiteral(THE_ISSUE_SEVERITY_EENUM, IssueSeverity[IssueSeverity.WARNING], IssueSeverity.WARNING);
-addLiteral(THE_ISSUE_SEVERITY_EENUM, IssueSeverity[IssueSeverity.ERROR], IssueSeverity.ERROR);
-
-export const THE_ISSUE_ECLASS = Ecore.EClass.create({
-    name: "Issue"
-});
-THE_ISSUE_ECLASS.get("eStructuralFeatures").add(Ecore.EAttribute.create({
-    name: "type",
-    eType: THE_ISSUE_TYPE_EENUM,
-    lowerBound: 1
-}));
-THE_ISSUE_ECLASS.get("eStructuralFeatures").add(Ecore.EAttribute.create({
-    name: "message",
-    eType: Ecore.EString,
-    lowerBound: 1
-}));
-THE_ISSUE_ECLASS.get("eStructuralFeatures").add(Ecore.EAttribute.create({
-    name: "severity",
-    eType: THE_ISSUE_SEVERITY_EENUM
-}));
-THE_ISSUE_ECLASS.get("eStructuralFeatures").add(Ecore.EReference.create({
-    name: "position",
-    eType: THE_POSITION_ECLASS,
-    containment: true
-}));
-
-export const THE_RESULT_ECLASS = Ecore.EClass.create({
-    name: "Result"
-});
-const resultTypeParameter = Ecore.ETypeParameter.create({ name: "CU" });
-(resultTypeParameter.get("eBounds") as EList).add(Ecore.EGenericType.create({
-    eClassifier: THE_NODE_ECLASS
-}));
-THE_RESULT_ECLASS.get("eTypeParameters").add(resultTypeParameter);
-THE_RESULT_ECLASS.get("eStructuralFeatures").add(Ecore.EReference.create({
-    name: "root",
-    eGenericType: Ecore.EGenericType.create({ eTypeParameter: resultTypeParameter }),
-    containment: true
-}));
-THE_RESULT_ECLASS.get("eStructuralFeatures").add(Ecore.EReference.create({
-    name: "issues",
-    eGenericType: THE_ISSUE_ECLASS,
-    containment: true,
-    upperBound: -1
-}));
-
-THE_AST_EPACKAGE.get('eClassifiers').add(THE_ORIGIN_ECLASS);
-THE_AST_EPACKAGE.get('eClassifiers').add(THE_NODE_ECLASS);
-THE_AST_EPACKAGE.get('eClassifiers').add(THE_POINT_ECLASS);
-THE_AST_EPACKAGE.get('eClassifiers').add(THE_POSITION_ECLASS);
-THE_AST_EPACKAGE.get('eClassifiers').add(THE_POSSIBLY_NAMED_INTERFACE);
-THE_AST_EPACKAGE.get('eClassifiers').add(THE_NAMED_INTERFACE);
-THE_AST_EPACKAGE.get('eClassifiers').add(THE_REFERENCE_BY_NAME_ECLASS);
-THE_AST_EPACKAGE.get('eClassifiers').add(THE_LOCAL_DATE_ECLASS);
-THE_AST_EPACKAGE.get('eClassifiers').add(THE_LOCAL_TIME_ECLASS);
-THE_AST_EPACKAGE.get('eClassifiers').add(THE_LOCAL_DATE_TIME_ECLASS);
-THE_AST_EPACKAGE.get('eClassifiers').add(THE_ISSUE_SEVERITY_EENUM);
-THE_AST_EPACKAGE.get('eClassifiers').add(THE_ISSUE_TYPE_EENUM);
-THE_AST_EPACKAGE.get('eClassifiers').add(THE_RESULT_ECLASS);
-
-export function getEPackage(packageName: string, args: { nsPrefix?: string; nsURI?: string }): EPackage {
-    const ePackage = Ecore.EPackage.Registry.ePackages().find(p => p.get("name") == packageName);
-    if(ePackage) {
-        if(args.nsURI && ePackage.get("nsURI") !== args.nsURI) {
-            throw new Error("Package " + packageName + " already exists with different nsUri: " + ePackage.get("nsURI"));
-        } else if(args.nsPrefix && ePackage.get("nsPrefix") !== args.nsPrefix) {
-            throw new Error("Package " + packageName + " already exists with different nsPrefix: " + ePackage.get("nsPrefix"));
-        } else {
-            return ePackage;
-        }
-    } else {
-        const newPackage = Ecore.EPackage.create({
-            name: packageName,
-            ...args
-        });
-        Ecore.EPackage.Registry.register(newPackage);
-        return newPackage;
-    }
-}
+import {
+    KOLASU_URI_V2, THE_AST_EPACKAGE,
+    THE_ISSUE_ECLASS, THE_ISSUE_SEVERITY_EENUM, THE_ISSUE_TYPE_EENUM,
+    THE_LOCAL_DATE_ECLASS,
+    THE_LOCAL_DATE_TIME_ECLASS,
+    THE_LOCAL_TIME_ECLASS,
+    THE_NODE_ECLASS, THE_POINT_ECLASS,
+    THE_POSITION_ECLASS,
+    THE_RESULT_ECLASS
+} from "./kolasu-v2-metamodel";
+import {KOLASU_URI_V1} from "./kolasu-v1-metamodel";
 
 function registerEPackage(packageName: string, args: { nsPrefix?: string; nsURI?: string }) {
     const packageDef = NODE_TYPES[packageName];
@@ -761,12 +511,12 @@ class ReferencesTracker {
  * @param data the input string or object.
  * @param resource where to look for to resolve references to types.
  */
-export function loadEObject(data: any, resource: Resource): EObject | undefined {
+export function loadEObject(data: any, resource: Resource, eClass?: EClass,): EObject | undefined {
     if(typeof data === "string") {
         data = JSON.parse(data);
     }
     const referencesTracker = new ReferencesTracker(resource);
-    const result = importJsonObject(data, resource, null, true, referencesTracker);
+    const result = importJsonObject(data, resource, eClass, true, referencesTracker);
     resource.add(result);
     referencesTracker.resolveAllReferences();
     return result;
@@ -778,7 +528,7 @@ export function findEClass(name: string, resource: Resource): EClass | undefined
         const packageName = name.substring(0, index);
         const ePackage = EPackage.Registry.getEPackage(packageName);
         if(!ePackage) {
-            throw new Error("Package not found: " + packageName);
+            throw new Error("Package not found: " + packageName+ " while loading for class " + name);
         }
         return ePackage.get("eClassifiers").find((c: EClassifier) =>
             c.get("name") == name.substring(name.lastIndexOf("/") + 1));
