+++ tests/data/parser-bench/rpgexamples/definition.json
@@ -1,0 +1,5 @@
+{
+  "key": "rpg",
+  "name": "RPG",
+  "description": "The RPG Language"
+}
